+++ core/src/main/kotlin/net/corda/core/flows/NotarisationRequest.kt
@@ -3,6 +3,7 @@ package net.corda.core.flows
 import net.corda.core.contracts.StateRef
 import net.corda.core.crypto.*
 import net.corda.core.identity.Party
+import net.corda.core.internal.notary.NotaryInternalException
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.serialization.serialize
 import net.corda.core.transactions.CoreTransaction
