+++ node/src/main/kotlin/net/corda/node/services/transactions/BFTNonValidatingNotaryService.kt
@@ -7,7 +7,8 @@ import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.SignedData
 import net.corda.core.flows.*
 import net.corda.core.identity.Party
-import net.corda.core.node.services.NotaryService
+import net.corda.core.internal.notary.NotaryInternalException
+import net.corda.core.internal.notary.NotaryService
 import net.corda.core.schemas.PersistentStateRef
 import net.corda.core.serialization.deserialize
 import net.corda.core.serialization.serialize
