+++ node/src/main/kotlin/net/corda/node/services/transactions/ValidatingNotaryService.kt
@@ -1,8 +1,8 @@
 package net.corda.node.services.transactions
 
 import net.corda.core.flows.FlowSession
-import net.corda.core.flows.NotaryFlow
-import net.corda.core.node.services.TrustedAuthorityNotaryService
+import net.corda.core.internal.notary.NotaryServiceFlow
+import net.corda.core.internal.notary.TrustedAuthorityNotaryService
 import net.corda.node.services.api.ServiceHubInternal
 import java.security.PublicKey
 
@@ -10,7 +10,7 @@ import java.security.PublicKey
 class ValidatingNotaryService(override val services: ServiceHubInternal, override val notaryIdentityKey: PublicKey) : TrustedAuthorityNotaryService() {
     override val uniquenessProvider = PersistentUniquenessProvider(services.clock)
 
-    override fun createServiceFlow(otherPartySession: FlowSession): NotaryFlow.Service = ValidatingNotaryFlow(otherPartySession, this)
+    override fun createServiceFlow(otherPartySession: FlowSession): NotaryServiceFlow = ValidatingNotaryFlow(otherPartySession, this)
 
     override fun start() {}
     override fun stop() {}
