+++ node/src/test/kotlin/net/corda/node/services/transactions/PersistentUniquenessProviderTests.kt
@@ -6,7 +6,7 @@ import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.sha256
 import net.corda.core.flows.NotarisationRequestSignature
 import net.corda.core.flows.NotaryError
-import net.corda.core.flows.NotaryInternalException
+import net.corda.core.internal.notary.NotaryInternalException
 import net.corda.core.identity.CordaX500Name
 import net.corda.node.internal.configureDatabase
 import net.corda.node.services.schema.NodeSchemaService
